# Optimize byte array encoding