--- a/encoding/delta/byte_array.go
+++ b/encoding/delta/byte_array.go
@@ -3,11 +3,14 @@ package delta
 import (
 	"bytes"
 	"fmt"
-	"math"
 
 	"github.com/segmentio/parquet-go/encoding"
 	"github.com/segmentio/parquet-go/encoding/plain"
 	"github.com/segmentio/parquet-go/format"
+)
+
+const (
+	maxLinearSearchPrefixLength = 128 // arbitrary
 )
 
 type ByteArrayEncoding struct {
@@ -23,27 +26,64 @@ func (e *ByteArrayEncoding) Encoding() format.Encoding {
 }
 
 func (e *ByteArrayEncoding) EncodeByteArray(dst, src []byte) ([]byte, error) {
-	lastOffset := int32(0)
+	prefix := getInt32Buffer()
+	defer putInt32Buffer(prefix)
 
-	offset := getInt32Buffer()
-	defer putInt32Buffer(offset)
+	length := getInt32Buffer()
+	defer putInt32Buffer(length)
 
-	err := plain.RangeByteArrays(src, func(value []byte) error {
-		offset.values = append(offset.values, lastOffset)
-		lastOffset += 4 + int32(len(value))
-		return nil
-	})
-	if err != nil {
-		return dst[:0], encoding.Error(e, err)
+	totalSize := 0
+	lastValue := ([]byte)(nil)
+
+	for i := 0; i < len(src); {
+		r := len(src) - i
+		if r < plain.ByteArrayLengthSize {
+			return dst[:0], plain.ErrTooShort(r)
+		}
+		n := plain.ByteArrayLength(src[i:])
+		i += plain.ByteArrayLengthSize
+		r -= plain.ByteArrayLengthSize
+		if n > r {
+			return dst[:0], plain.ErrTooShort(n)
+		}
+		if n > plain.MaxByteArrayLength {
+			return dst[:0], plain.ErrTooLarge(n)
+		}
+		v := src[i : i+n : i+n]
+		p := 0
+
+		if len(v) <= maxLinearSearchPrefixLength {
+			p = linearSearchPrefixLength(lastValue, v)
+		} else {
+			p = binarySearchPrefixLength(len(lastValue)/2, lastValue, v)
+		}
+
+		prefix.values = append(prefix.values, int32(p))
+		length.values = append(length.values, int32(n-p))
+		lastValue = v
+		totalSize += n - p
+		i += n
 	}
 
-	return e.encode(dst[:0], len(offset.values), func(i int) []byte {
-		j := int(offset.values[i])
-		k := j + plain.ByteArrayLength(src[j:])
-		j += 4
-		k += 4
-		return src[j:k:k]
-	})
+	dst = dst[:0]
+	dst = encodeInt32(dst, prefix.values)
+	dst = encodeInt32(dst, length.values)
+	dst = resize(dst, len(dst)+totalSize)
+
+	b := dst[len(dst)-totalSize:]
+	i := plain.ByteArrayLengthSize
+	j := 0
+
+	_ = length.values[:len(prefix.values)]
+
+	for k, p := range prefix.values {
+		n := p + length.values[k]
+		j += copy(b[j:], src[i+int(p):i+int(n)])
+		i += plain.ByteArrayLengthSize
+		i += int(n)
+	}
+
+	return dst, nil
 }
 
 func (e *ByteArrayEncoding) EncodeFixedLenByteArray(dst, src []byte, size int) ([]byte, error) {
@@ -57,37 +97,40 @@ func (e *ByteArrayEncoding) EncodeFixedLenByteArray(dst, src []byte, size int) (
 	if (len(src) % size) != 0 {
 		return dst[:0], encoding.ErrEncodeInvalidInputSize(e, "FIXED_LEN_BYTE_ARRAY", len(src))
 	}
-	return e.encode(dst[:0], len(src)/size, func(i int) []byte {
-		j := (i + 0) * size
-		k := (i + 1) * size
-		return src[j:k:k]
-	})
-}
 
-func (e *ByteArrayEncoding) encode(dst []byte, numValues int, valueAt func(int) []byte) ([]byte, error) {
 	prefix := getInt32Buffer()
 	defer putInt32Buffer(prefix)
 
 	length := getInt32Buffer()
 	defer putInt32Buffer(length)
 
-	var lastValue []byte
-	for i := 0; i < numValues; i++ {
-		value := valueAt(i)
-		if len(value) > math.MaxInt32 {
-			return dst, encoding.Errorf(e, "byte array of length %d is too large to be encoded", len(value))
-		}
-		n := prefixLength(lastValue, value)
-		prefix.values = append(prefix.values, int32(n))
-		length.values = append(length.values, int32(len(value)-n))
-		lastValue = value
+	totalSize := 0
+	lastValue := ([]byte)(nil)
+
+	for i := size; i <= len(src); i += size {
+		v := src[i-size : i : i]
+		p := linearSearchPrefixLength(lastValue, v)
+		n := size - p
+		prefix.values = append(prefix.values, int32(p))
+		length.values = append(length.values, int32(n))
+		lastValue = v
+		totalSize += n
 	}
 
+	dst = dst[:0]
 	dst = encodeInt32(dst, prefix.values)
 	dst = encodeInt32(dst, length.values)
-	for i, p := range prefix.values {
-		dst = append(dst, valueAt(i)[p:]...)
+	dst = resize(dst, len(dst)+totalSize)
+
+	b := dst[len(dst)-totalSize:]
+	i := 0
+	j := 0
+
+	for _, p := range prefix.values {
+		j += copy(b[j:], src[i+int(p):i+size])
+		i += size
 	}
+
 	return dst, nil
 }
 
@@ -179,8 +222,11 @@ func (e *ByteArrayEncoding) decode(src []byte, observe func(prefix, suffix []byt
 	return nil
 }
 
-func prefixLength(base, data []byte) int {
-	return binarySearchPrefixLength(len(base)/2, base, data)
+func linearSearchPrefixLength(base, data []byte) (n int) {
+	for n < len(base) && n < len(data) && base[n] == data[n] {
+		n++
+	}
+	return n
 }
 
 func binarySearchPrefixLength(max int, base, data []byte) int {