--- a/encoding/plain/plain_amd64.go
+++ b/encoding/plain/plain_amd64.go
@@ -1,0 +1,6 @@
+//go:build !purego
+
+package plain
+
+//go:noescape
+func validateByteArrays([]byte) status