--- a/encoding/plain/plain_test.go
+++ b/encoding/plain/plain_test.go
@@ -32,3 +32,27 @@ func TestAppendBoolean(t *testing.T) {
 		t.Errorf("%08b\n", values)
 	}
 }
+
+func TestValidateByteArrays(t *testing.T) {
+	t.Run("ok", func(t *testing.T) {
+		var b []byte
+		b = plain.AppendByteArrayString(b, "Hello")
+		b = plain.AppendByteArrayString(b, "World")
+		b = plain.AppendByteArrayString(b, "!")
+
+		if err := plain.ValidateByteArrays(b); err != nil {
+			t.Error(err)
+		}
+	})
+
+	t.Run("errTooShort", func(t *testing.T) {
+		var b []byte
+		b = plain.AppendByteArrayString(b, "Hello")
+		b = plain.AppendByteArrayString(b, "World")
+		b = plain.AppendByteArrayString(b, "!")
+
+		if plain.ValidateByteArrays(b[:len(b)-1]) == nil {
+			t.Error("expected non-nil error")
+		}
+	})
+}