… encoding/plain/plain_amd64.s