Follow up to https://github.com/segmentio/parquet-go/pull/200, this PR improves the throughput of byte array encodings. These are important code paths because they are involved when encoding string values, which are a very common case.

Most of these optimizations are simply refactors of the Go code to reduce the number of passes on the inputs, and avoid function calls on the busy code paths.

```
name                                             old time/op    new time/op    delta
Encode/PLAIN/byte_array                            63.7µs ± 0%    26.8µs ± 0%   -57.86%  (p=0.000 n=9+9)
Encode/PLAIN/fixed_length_byte_array               3.16µs ± 0%    3.16µs ± 0%    +0.10%  (p=0.030 n=8+10)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array           182µs ± 0%     106µs ± 1%   -41.75%  (p=0.000 n=9+9)
Encode/DELTA_BYTE_ARRAY/byte_array                  445µs ± 0%     141µs ± 1%   -68.36%  (p=0.000 n=10+9)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array     279µs ± 1%     103µs ± 0%   -62.94%  (p=0.000 n=9+10)

name                                             old speed      new speed      delta
Encode/PLAIN/byte_array                          2.27GB/s ± 0%  5.39GB/s ± 0%  +137.28%  (p=0.000 n=9+9)
Encode/PLAIN/fixed_length_byte_array             50.6GB/s ± 0%  50.6GB/s ± 0%    -0.11%  (p=0.017 n=9+10)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array         793MB/s ± 0%  1361MB/s ± 1%   +71.68%  (p=0.000 n=9+9)
Encode/DELTA_BYTE_ARRAY/byte_array                325MB/s ± 0%  1028MB/s ± 1%  +216.03%  (p=0.000 n=10+9)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array   574MB/s ± 1%  1549MB/s ± 0%  +169.84%  (p=0.000 n=9+10)

name                                             old value/s    new value/s    delta
Encode/PLAIN/byte_array                              157M ± 0%      373M ± 0%  +137.28%  (p=0.000 n=9+9)
Encode/PLAIN/fixed_length_byte_array                3.16G ± 0%     3.16G ± 0%    -0.11%  (p=0.017 n=9+10)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array           54.8M ± 0%     94.1M ± 1%   +71.68%  (p=0.000 n=9+9)
Encode/DELTA_BYTE_ARRAY/byte_array                  22.5M ± 0%     71.1M ± 1%  +216.04%  (p=0.000 n=10+9)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array     35.9M ± 1%     96.8M ± 0%  +169.84%  (p=0.000 n=9+10)
```